--- photo-sharing-client/src/Components/AuthForm.js
@@ -1,113 +1,0 @@
-// AuthForm.js
-
-import React from "react";
-import Button from "@mui/material/Button";
-import TextField from "@mui/material/TextField";
-import Box from "@mui/material/Box";
-import { Link } from "react-router-dom";
-
-import { useState } from "react";
-import { initializeApp } from "firebase/app";
-import {
-  getAuth,
-  createUserWithEmailAndPassword,
-  signInWithEmailAndPassword,
-} from "firebase/auth";
-
-const firebaseConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
-  authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.REACT_APP_PROJECT_ID,
-  storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.REACT_APP_FIREBASE_API_ID,
-};
-
-const app = initializeApp(firebaseConfig);
-const auth = getAuth(app);
-
-const AuthForm = ({ isSignup, setIsLogged }) => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [name, setName] = useState("");
-
-  const handleSubmit = async (event) => {
-    event.preventDefault();
-
-    if (isSignup) {
-      createUserWithEmailAndPassword(auth, email, password)
-        .then((userCredential) => {
-          setIsLogged(true);
-          alert("User created successfully");
-          setEmail("");
-          setPassword("");
-          setName("");
-        })
-        .catch((error) => {
-          alert(error.message);
-        });
-    } else {
-      signInWithEmailAndPassword(auth, email, password)
-        .then((userCredential) => {
-          setIsLogged(true);
-          alert("Logged in successfully");
-          setEmail("");
-          setPassword("");
-        })
-        .catch((error) => {
-          alert(error.message);
-        });
-    }
-  };
-
-  return (
-    <Box
-      component="form"
-      onSubmit={handleSubmit}
-      sx={{
-        "& > :not(style)": { m: 1, width: "25ch" },
-      }}
-      noValidate
-      autoComplete="off"
-    >
-      {isSignup && (
-        <TextField
-          id="outlined-basic"
-          label="Name"
-          variant="outlined"
-          value={name}
-          onChange={(e) => setName(e.target.value)}
-        />
-      )}
-      <TextField
-        id="outlined-basic"
-        label="Email"
-        variant="outlined"
-        value={email}
-        onChange={(e) => setEmail(e.target.value)}
-      />
-      <TextField
-        id="outlined-basic"
-        label="Password"
-        variant="outlined"
-        type="password"
-        value={password}
-        onChange={(e) => setPassword(e.target.value)}
-      />
-      <Button variant="contained" type="submit">
-        {isSignup ? "Sign Up" : "Log In"}
-      </Button>
-      {isSignup ? (
-        <p>
-          Already have an account? <Link to="/login">Log in here</Link>
-        </p>
-      ) : (
-        <p>
-          Don't have an account? <Link to="/signup">Sign up here</Link>
-        </p>
-      )}
-    </Box>
-  );
-};
-
-export default AuthForm;
